refactor: tests for `executeRelayCall(bytes,uint256,bytes)` -> `execute(bytes)`
