+++ tests/LSP6KeyManager/tests/Reentrancy.test.ts
@@ -1,150 +1,46 @@
-import { expect } from "chai";
-import { ethers } from "hardhat";
-
-//types
-import { SignerWithAddress } from "@nomiclabs/hardhat-ethers/signers";
-import { BytesLike, Wallet } from "ethers";
-import {
-  ReentrantContract__factory,
-  ReentrantContract,
-  UniversalProfile__factory,
-} from "../../../types";
-
-// constants
-import {
-  ERC725YDataKeys,
-  ALL_PERMISSIONS,
-  PERMISSIONS,
-} from "../../../constants";
-
 // setup
 import { LSP6TestContext } from "../../utils/context";
+import { buildReentrancyContext } from "./Reentrancy/reentrancyHelpers";
 
-// helpers
-import {
-  encodeCompactBytesArray,
-  combinePermissions,
-  combineAllowedCalls,
-  LOCAL_PRIVATE_KEYS,
-} from "../../utils/helpers";
-import {
-  ReentrancyContext,
-  buildReentrancyContext,
-  testCasesByType,
-  transferValueTestCases,
-  setDataTestCases,
-  addPermissionsTestCases,
-  changePermissionsTestCases,
-  addUniversalReceiverDelegateTestCases,
-  changeUniversalReceiverDelegateTestCases,
-} from "./reentrancyHelpers";
+// tests
+import { testSingleExecuteToSingleExecute } from "./Reentrancy/SingleExecuteToSingleExecute.test";
+import { testSingleExecuteRelayCallToSingleExecute } from "./Reentrancy/SingleExecuteRelayCallToSingleExecute.test";
+import { testSingleExecuteToSingleExecuteRelayCall } from "./Reentrancy/SingleExecuteToSingleExecuteRelayCall.test";
+import { testSingleExecuteRelayCallToSingleExecuteRelayCall } from "./Reentrancy/SingleExecuteRelayCallToSingleExecuteRelayCall.test";
 
 export const testReentrancyScenarios = (
   buildContext: () => Promise<LSP6TestContext>
 ) => {
-  let context: LSP6TestContext;
-  let reentrancyContext: ReentrancyContext;
-
-  before(async () => {
-    context = await buildContext();
-    reentrancyContext = await buildReentrancyContext(context);
-  });
-
   describe("first call through `execute(bytes)`, second call through `execute(bytes)`", () => {
-    describe("when reentering and transferring value", () => {
-      transferValueTestCases.forEach((testCase) => {
-        it(`${testCase.testDescription}`, async () => {
-          await testCasesByType(
-            "TRANSFERVALUE",
-            testCase,
-            context,
-            reentrancyContext
-          );
-        });
-      });
-    });
-
-    describe("when reentering and setting data", () => {
-      setDataTestCases.forEach((testCase) => {
-        it(`${testCase.testDescription}`, async () => {
-          await testCasesByType(
-            "SETDATA",
-            testCase,
-            context,
-            reentrancyContext
-          );
-        });
-      });
-    });
-
-    describe("when reentering and adding permissions", () => {
-      addPermissionsTestCases.forEach((testCase) => {
-        it(`${testCase.testDescription}`, async () => {
-          await testCasesByType(
-            "ADDPERMISSIONS",
-            testCase,
-            context,
-            reentrancyContext
-          );
-        });
-      });
-    });
-
-    describe("when reentering and changing permissions", () => {
-      changePermissionsTestCases.forEach((testCase) => {
-        it(`${testCase.testDescription}`, async () => {
-          await testCasesByType(
-            "CHANGEPERMISSIONS",
-            testCase,
-            context,
-            reentrancyContext
-          );
-        });
-      });
-    });
-
-    describe("when reentering and adding URD", () => {
-      addUniversalReceiverDelegateTestCases.forEach((testCase) => {
-        it(`${testCase.testDescription}`, async () => {
-          await testCasesByType(
-            "ADDUNIVERSALRECEIVERDELEGATE",
-            testCase,
-            context,
-            reentrancyContext
-          );
-        });
-      });
-    });
-
-    describe("when reentering and changing URD", () => {
-      changeUniversalReceiverDelegateTestCases.forEach((testCase) => {
-        it(`${testCase.testDescription}`, async () => {
-          await testCasesByType(
-            "CHANGEUNIVERSALRECEIVERDELEGATE",
-            testCase,
-            context,
-            reentrancyContext
-          );
-        });
-      });
-    });
-
-    after(async () => {
-      await reentrancyContext.owner.sendTransaction({
-        to: context.universalProfile.address,
-        value: ethers.utils.parseEther("1"),
-      });
-    });
+    testSingleExecuteToSingleExecute(buildContext, buildReentrancyContext);
   });
 
   describe("first call through `executeRelayCall(bytes,uint256,bytes)`, second call through `execute(bytes)`", () => {
-    describe("when reentering and transferring value", () => {});
+    testSingleExecuteRelayCallToSingleExecute(
+      buildContext,
+      buildReentrancyContext
+    );
+  });
 
-    after(async () => {
-      await reentrancyContext.owner.sendTransaction({
-        to: context.universalProfile.address,
-        value: ethers.utils.parseEther("1"),
-      });
-    });
+  describe.only("first call through `execute(bytes)`, second call through `executeRelayCall(bytes,uint256,bytes)`", () => {
+    testSingleExecuteToSingleExecuteRelayCall(
+      buildContext,
+      buildReentrancyContext
+    );
+  });
+
+  describe("first call through `executeRelayCall(bytes,uint256,bytes)`, second call through `executeRelayCall(bytes,uint256,bytes)`", () => {
+    testSingleExecuteRelayCallToSingleExecuteRelayCall(
+      buildContext,
+      buildReentrancyContext
+    );
+  });
+
+  describe("first call through `execute(bytes)`, second call through `execute(uint256[],bytes[])`", () => {
+    //testSingleExecuteToBatchExecute(buildContext, buildReentrancyContext);
+  });
+
+  describe("first call through `execute(bytes)`, second call through `executeRelayCall(bytes[],uint256[],uint256[],bytes[])`", () => {
+    //testSingleExecuteToBatchExecuteRelayCall(buildContext, buildReentrancyContext);
   });
 };
