+++ tests/LSP6KeyManager/tests/Reentrancy/reentrancyHelpers.ts
@@ -1,32 +1,34 @@
-import { expect } from "chai";
 import { ethers } from "hardhat";
 
 //types
 import { SignerWithAddress } from "@nomiclabs/hardhat-ethers/signers";
-import { BytesLike, Wallet } from "ethers";
+import { BigNumber, BytesLike, Wallet } from "ethers";
 import {
   ReentrantContract__factory,
   ReentrantContract,
-  UniversalProfile__factory,
-} from "../../../types";
+  LSP6KeyManager,
+  UniversalProfile,
+  RelayBatchReentrancy,
+  RelaySingleReentrancy,
+} from "../../../../types";
 
 // constants
 import {
   ERC725YDataKeys,
   ALL_PERMISSIONS,
   PERMISSIONS,
-} from "../../../constants";
+} from "../../../../constants";
 
 // setup
-import { LSP6TestContext } from "../../utils/context";
+import { LSP6TestContext } from "../../../utils/context";
 
 // helpers
 import {
-  encodeCompactBytesArray,
   combinePermissions,
   combineAllowedCalls,
   LOCAL_PRIVATE_KEYS,
-} from "../../utils/helpers";
+  signLSP6ExecuteRelayCall,
+} from "../../../utils/helpers";
 
 // Complex permission as it has AllowedCalls
 export type TransferValueTestCase = {
@@ -65,372 +67,8 @@ export type ReentrancyContext = {
   caller: SignerWithAddress;
   signer: Wallet;
   reentrantContract: ReentrantContract;
+  reentrantSigner: Wallet;
   randomLSP1TypeId: string;
-};
-
-export const buildReentrancyContext = async (context: LSP6TestContext) => {
-  const owner = context.accounts[7];
-  const caller = context.accounts[8];
-  const signer = new ethers.Wallet(LOCAL_PRIVATE_KEYS.ACCOUNT7);
-
-  const reentrantContract = await await new ReentrantContract__factory(
-    owner
-  ).deploy(
-    context.accounts[9].address,
-    ethers.utils.keccak256(ethers.utils.toUtf8Bytes("RandomLSP1TypeId")),
-    context.accounts[9].address
-  );
-
-  const permissionKeys = [
-    ERC725YDataKeys.LSP6["AddressPermissions:Permissions"] +
-      owner.address.substring(2),
-    ERC725YDataKeys.LSP6["AddressPermissions:Permissions"] +
-      caller.address.substring(2),
-    ERC725YDataKeys.LSP6["AddressPermissions:AllowedCalls"] +
-      caller.address.substring(2),
-    ERC725YDataKeys.LSP6["AddressPermissions:Permissions"] +
-      signer.address.substring(2),
-  ];
-
-  const permissionValues = [
-    ALL_PERMISSIONS,
-    PERMISSIONS.CALL,
-    combineAllowedCalls(
-      ["0xffffffff"],
-      [reentrantContract.address],
-      ["0xffffffff"]
-    ),
-    PERMISSIONS.CALL,
-  ];
-
-  await context.universalProfile
-    .connect(context.accounts[0])
-    ["setData(bytes32[],bytes[])"](permissionKeys, permissionValues);
-
-  await context.universalProfile
-    .connect(context.accounts[0])
-    .transferOwnership(context.keyManager.address);
-
-  const acceptOwnershipPayload =
-    context.universalProfile.interface.encodeFunctionData("acceptOwnership");
-  await context.keyManager
-    .connect(owner)
-    ["execute(bytes)"](acceptOwnershipPayload);
-
-  // Fund Universal Profile with some LYXe
-  await owner.sendTransaction({
-    to: context.universalProfile.address,
-    value: ethers.utils.parseEther("10"),
-  });
-
-  const randomLSP1TypeId = ethers.utils.keccak256(
-    ethers.utils.toUtf8Bytes("RandomLSP1TypeId")
-  );
-
-  return {
-    owner,
-    caller,
-    signer,
-    reentrantContract,
-    randomLSP1TypeId,
-  };
-};
-
-const loadTestCase = async (
-  payloadType: string,
-  testCase: TransferValueTestCase | SetDataTestCase | SimplePermissionTestCase,
-  context: LSP6TestContext,
-  reentrancyContext: ReentrancyContext
-) => {
-  let permissionKeys: BytesLike[];
-  let permissionValues: BytesLike[];
-
-  switch (payloadType) {
-    case "TRANSFERVALUE": {
-      permissionKeys = [
-        ERC725YDataKeys.LSP6["AddressPermissions:Permissions"] +
-          reentrancyContext.reentrantContract.address.substring(2),
-        ERC725YDataKeys.LSP6["AddressPermissions:AllowedCalls"] +
-          reentrancyContext.reentrantContract.address.substring(2),
-      ];
-
-      permissionValues = [
-        testCase.permissions,
-        (testCase as TransferValueTestCase).allowedCalls
-          ? combineAllowedCalls(
-              ["0xffffffff"],
-              [reentrancyContext.reentrantContract.address],
-              ["0xffffffff"]
-            )
-          : "0x",
-      ];
-      break;
-    }
-    case "SETDATA": {
-      permissionKeys = [
-        ERC725YDataKeys.LSP6["AddressPermissions:Permissions"] +
-          reentrancyContext.reentrantContract.address.substring(2),
-        ERC725YDataKeys.LSP6["AddressPermissions:AllowedERC725YDataKeys"] +
-          reentrancyContext.reentrantContract.address.substring(2),
-      ];
-
-      permissionValues = [
-        testCase.permissions,
-        (testCase as SetDataTestCase).allowedERC725YDataKeys
-          ? encodeCompactBytesArray([
-              ethers.utils.keccak256(
-                ethers.utils.toUtf8Bytes("SomeRandomTextUsed")
-              ),
-            ])
-          : "0x",
-      ];
-      break;
-    }
-    default: {
-      permissionKeys = [
-        ERC725YDataKeys.LSP6["AddressPermissions:Permissions"] +
-          reentrancyContext.reentrantContract.address.substring(2),
-      ];
-
-      permissionValues = [testCase.permissions];
-    }
-  }
-
-  const permissionsPayload =
-    new UniversalProfile__factory().interface.encodeFunctionData(
-      "setData(bytes32[],bytes[])",
-      [permissionKeys, permissionValues]
-    );
-  await context.keyManager
-    .connect(reentrancyContext.owner)
-    ["execute(bytes)"](permissionsPayload);
-};
-
-const testNotAuthorisedErrorCase = async (
-  payloadType: string,
-  testCase: TransferValueTestCase | SetDataTestCase | SimplePermissionTestCase,
-  context: LSP6TestContext,
-  reentrancyContext: ReentrancyContext
-) => {
-  const reentrantPayload =
-    new ReentrantContract__factory().interface.encodeFunctionData(
-      "callThatReenters",
-      [context.keyManager.address, payloadType]
-    );
-
-  const executePayload =
-    new UniversalProfile__factory().interface.encodeFunctionData(
-      "execute(uint256,address,uint256,bytes)",
-      [0, reentrancyContext.reentrantContract.address, 0, reentrantPayload]
-    );
-
-  await expect(
-    context.keyManager
-      .connect(reentrancyContext.caller)
-      ["execute(bytes)"](executePayload)
-  )
-    .to.be.revertedWithCustomError(context.keyManager, testCase.customErrorName)
-    .withArgs(
-      reentrancyContext.reentrantContract.address,
-      testCase.customErrorArgs.permission
-    );
-};
-
-const testEmptyCustomErrorCase = async (
-  payloadType: string,
-  testCase: TransferValueTestCase | SetDataTestCase,
-  context: LSP6TestContext,
-  reentrancyContext: ReentrancyContext
-) => {
-  const reentrantPayload =
-    new ReentrantContract__factory().interface.encodeFunctionData(
-      "callThatReenters",
-      [context.keyManager.address, payloadType]
-    );
-
-  const executePayload =
-    new UniversalProfile__factory().interface.encodeFunctionData(
-      "execute(uint256,address,uint256,bytes)",
-      [0, reentrancyContext.reentrantContract.address, 0, reentrantPayload]
-    );
-
-  await expect(
-    context.keyManager
-      .connect(reentrancyContext.caller)
-      ["execute(bytes)"](executePayload)
-  ).to.be.revertedWithCustomError(context.keyManager, testCase.customErrorName);
-};
-
-const testValidCase = async (
-  payloadType: string,
-  context: LSP6TestContext,
-  reentrancyContext: ReentrancyContext
-) => {
-  const reentrantPayload =
-    new ReentrantContract__factory().interface.encodeFunctionData(
-      "callThatReenters",
-      [context.keyManager.address, payloadType]
-    );
-
-  const executePayload =
-    new UniversalProfile__factory().interface.encodeFunctionData(
-      "execute(uint256,address,uint256,bytes)",
-      [0, reentrancyContext.reentrantContract.address, 0, reentrantPayload]
-    );
-
-  switch (payloadType) {
-    case "TRANSFERVALUE": {
-      expect(
-        await context.universalProfile.provider.getBalance(
-          context.universalProfile.address
-        )
-      ).to.equal(ethers.utils.parseEther("10"));
-
-      await context.keyManager
-        .connect(reentrancyContext.caller)
-        ["execute(bytes)"](executePayload);
-
-      expect(
-        await context.universalProfile.provider.getBalance(
-          context.universalProfile.address
-        )
-      ).to.equal(ethers.utils.parseEther("9"));
-
-      expect(
-        await context.universalProfile.provider.getBalance(
-          reentrancyContext.reentrantContract.address
-        )
-      ).to.equal(ethers.utils.parseEther("1"));
-      break;
-    }
-    case "SETDATA": {
-      await context.keyManager
-        .connect(reentrancyContext.caller)
-        ["execute(bytes)"](executePayload);
-
-      const hardcodedKey = ethers.utils.keccak256(
-        ethers.utils.toUtf8Bytes("SomeRandomTextUsed")
-      );
-      const hardcodedValue = ethers.utils.hexlify(
-        ethers.utils.toUtf8Bytes("SomeRandomTextUsed")
-      );
-
-      expect(
-        await context.universalProfile["getData(bytes32)"](hardcodedKey)
-      ).to.equal(hardcodedValue);
-      break;
-    }
-    case "ADDPERMISSIONS": {
-      await context.keyManager
-        .connect(reentrancyContext.caller)
-        ["execute(bytes)"](executePayload);
-
-      const hardcodedPermissionKey =
-        ERC725YDataKeys.LSP6["AddressPermissions:Permissions"] +
-        context.accounts[9].address.substring(2);
-      const hardcodedPermissionValue =
-        "0x0000000000000000000000000000000000000000000000000000000000000010";
-
-      expect(
-        await context.universalProfile["getData(bytes32)"](
-          hardcodedPermissionKey
-        )
-      ).to.equal(hardcodedPermissionValue);
-      break;
-    }
-    case "CHANGEPERMISSIONS": {
-      await context.keyManager
-        .connect(reentrancyContext.caller)
-        ["execute(bytes)"](executePayload);
-
-      const hardcodedPermissionKey =
-        ERC725YDataKeys.LSP6["AddressPermissions:Permissions"] +
-        context.accounts[9].address.substring(2);
-      const hardcodedPermissionValue = "0x";
-
-      expect(
-        await context.universalProfile["getData(bytes32)"](
-          hardcodedPermissionKey
-        )
-      ).to.equal(hardcodedPermissionValue);
-      break;
-    }
-    case "ADDUNIVERSALRECEIVERDELEGATE": {
-      await context.keyManager
-        .connect(reentrancyContext.caller)
-        ["execute(bytes)"](executePayload);
-
-      const hardcodedLSP1Key =
-        ERC725YDataKeys.LSP1.LSP1UniversalReceiverDelegatePrefix +
-        reentrancyContext.randomLSP1TypeId.substring(2, 42);
-
-      const hardcodedLSP1Value = context.accounts[9].address;
-
-      expect(
-        await context.universalProfile["getData(bytes32)"](hardcodedLSP1Key)
-      ).to.equal(hardcodedLSP1Value);
-      break;
-    }
-    case "CHANGEUNIVERSALRECEIVERDELEGATE": {
-      await context.keyManager
-        .connect(reentrancyContext.caller)
-        ["execute(bytes)"](executePayload);
-
-      const hardcodedLSP1Key =
-        ERC725YDataKeys.LSP1.LSP1UniversalReceiverDelegatePrefix +
-        reentrancyContext.randomLSP1TypeId.substring(2, 42);
-
-      const hardcodedLSP1Value = "0x";
-
-      expect(
-        await context.universalProfile["getData(bytes32)"](hardcodedLSP1Key)
-      ).to.equal(hardcodedLSP1Value);
-      break;
-    }
-  }
-};
-
-export const testCasesByType = async (
-  payloadType: string,
-  testCase: TransferValueTestCase | SetDataTestCase | SimplePermissionTestCase,
-  context: LSP6TestContext,
-  reentrancyContext: ReentrancyContext
-) => {
-  await loadTestCase(payloadType, testCase, context, reentrancyContext);
-
-  if (testCase.customErrorName == "NotAuthorised")
-    await testNotAuthorisedErrorCase(
-      payloadType,
-      testCase,
-      context,
-      reentrancyContext
-    );
-  else if (
-    testCase.customErrorName == "NoCallsAllowed" ||
-    testCase.customErrorName == "NoERC725YDataKeysAllowed"
-  )
-    switch (payloadType) {
-      case "TRANSFERVALUE": {
-        await testEmptyCustomErrorCase(
-          payloadType,
-          testCase as TransferValueTestCase,
-          context,
-          reentrancyContext
-        );
-        break;
-      }
-      case "SETDATA": {
-        await testEmptyCustomErrorCase(
-          payloadType,
-          testCase as SetDataTestCase,
-          context,
-          reentrancyContext
-        );
-        break;
-      }
-    }
-  else if (testCase.customErrorName == "")
-    await testValidCase(payloadType, context, reentrancyContext);
 };
 
 export const transferValueTestCases: TransferValueTestCase[] = [
@@ -816,3 +454,292 @@ export const changeUniversalReceiverDelegateTestCases: SimplePermissionTestCase[
       },
     },
   ];
+
+export const buildReentrancyContext = async (context: LSP6TestContext) => {
+  const owner = context.accounts[0];
+  const caller = context.accounts[1];
+  const signer = new ethers.Wallet(LOCAL_PRIVATE_KEYS.ACCOUNT2);
+
+  const reentrantContract = await await new ReentrantContract__factory(
+    owner
+  ).deploy(
+    context.accounts[9].address,
+    ethers.utils.keccak256(ethers.utils.toUtf8Bytes("RandomLSP1TypeId")),
+    context.accounts[9].address
+  );
+
+  const reentrantSigner = new ethers.Wallet(LOCAL_PRIVATE_KEYS.ACCOUNT3);
+
+  const permissionKeys = [
+    ERC725YDataKeys.LSP6["AddressPermissions:Permissions"] +
+      owner.address.substring(2),
+    ERC725YDataKeys.LSP6["AddressPermissions:Permissions"] +
+      caller.address.substring(2),
+    ERC725YDataKeys.LSP6["AddressPermissions:AllowedCalls"] +
+      caller.address.substring(2),
+    ERC725YDataKeys.LSP6["AddressPermissions:Permissions"] +
+      signer.address.substring(2),
+    ERC725YDataKeys.LSP6["AddressPermissions:AllowedCalls"] +
+      signer.address.substring(2),
+  ];
+
+  const permissionValues = [
+    ALL_PERMISSIONS,
+    PERMISSIONS.CALL,
+    combineAllowedCalls(
+      ["0xffffffff"],
+      [reentrantContract.address],
+      ["0xffffffff"]
+    ),
+    PERMISSIONS.CALL,
+    combineAllowedCalls(
+      ["0xffffffff"],
+      [reentrantContract.address],
+      ["0xffffffff"]
+    ),
+  ];
+
+  await context.universalProfile
+    .connect(context.accounts[0])
+    ["setData(bytes32[],bytes[])"](permissionKeys, permissionValues);
+
+  await context.universalProfile
+    .connect(context.accounts[0])
+    .transferOwnership(context.keyManager.address);
+
+  const acceptOwnershipPayload =
+    context.universalProfile.interface.encodeFunctionData("acceptOwnership");
+  await context.keyManager
+    .connect(owner)
+    ["execute(bytes)"](acceptOwnershipPayload);
+
+  // Fund Universal Profile with some LYXe
+  await owner.sendTransaction({
+    to: context.universalProfile.address,
+    value: ethers.utils.parseEther("10"),
+  });
+
+  const randomLSP1TypeId = ethers.utils.keccak256(
+    ethers.utils.toUtf8Bytes("RandomLSP1TypeId")
+  );
+
+  return {
+    owner,
+    caller,
+    signer,
+    reentrantContract,
+    reentrantSigner,
+    randomLSP1TypeId,
+  };
+};
+
+export const generateRelayCall = async (
+  keyManager: LSP6KeyManager,
+  payload: BytesLike,
+  signer: Wallet
+) => {
+  let nonce = await keyManager.callStatic.getNonce(signer.address, 1);
+  let msgValue = 0;
+  let signature = await signLSP6ExecuteRelayCall(
+    keyManager,
+    nonce.toString(),
+    signer.privateKey,
+    msgValue,
+    payload.toString()
+  );
+
+  const relayCallContext: {
+    signature: BytesLike;
+    nonce: BigNumber;
+    payload: BytesLike;
+  } = {
+    signature,
+    nonce,
+    payload,
+  };
+
+  return relayCallContext;
+};
+
+export const generateSingleRelayPayload = async (
+  reentrancyRelayer: RelaySingleReentrancy,
+  universalProfile: UniversalProfile,
+  keyManager: LSP6KeyManager,
+  reentrantSigner: Wallet,
+  payloadType: string,
+  testedAddress: string
+) => {
+  let payload: BytesLike;
+  switch (payloadType) {
+    case "TRANSFERVALUE":
+      payload = universalProfile.interface.encodeFunctionData(
+        "execute(uint256,address,uint256,bytes)",
+        [0, reentrancyRelayer.address, ethers.utils.parseEther("1"), "0x"]
+      );
+      break;
+    case "SETDATA":
+      payload = universalProfile.interface.encodeFunctionData(
+        "setData(bytes32,bytes)",
+        [
+          ethers.utils.keccak256(
+            ethers.utils.toUtf8Bytes("SomeRandomTextUsed")
+          ),
+          ethers.utils.hexlify(ethers.utils.toUtf8Bytes("SomeRandomTextUsed")),
+        ]
+      );
+      break;
+    case "ADDPERMISSIONS":
+      payload = universalProfile.interface.encodeFunctionData(
+        "setData(bytes32,bytes)",
+        [
+          ERC725YDataKeys.LSP6["AddressPermissions:Permissions"] +
+            testedAddress.substring(2),
+          ALL_PERMISSIONS,
+        ]
+      );
+      break;
+    case "CHANGEPERMISSIONS":
+      payload = universalProfile.interface.encodeFunctionData(
+        "setData(bytes32,bytes)",
+        [
+          ERC725YDataKeys.LSP6["AddressPermissions:Permissions"] +
+            testedAddress.substring(2),
+          "0x",
+        ]
+      );
+      break;
+    case "ADDUNIVERSALRECEIVERDELEGATE":
+      payload = universalProfile.interface.encodeFunctionData(
+        "setData(bytes32,bytes)",
+        [
+          ERC725YDataKeys.LSP1.LSP1UniversalReceiverDelegatePrefix +
+            ethers.utils
+              .keccak256(ethers.utils.toUtf8Bytes("RandomLSP1TypeId"))
+              .substring(2, 42),
+          testedAddress,
+        ]
+      );
+      break;
+    case "CHANGEUNIVERSALRECEIVERDELEGATE":
+      payload = universalProfile.interface.encodeFunctionData(
+        "setData(bytes32,bytes)",
+        [
+          ERC725YDataKeys.LSP1.LSP1UniversalReceiverDelegatePrefix +
+            ethers.utils
+              .keccak256(ethers.utils.toUtf8Bytes("RandomLSP1TypeId"))
+              .substring(2, 42),
+          "0x",
+        ]
+      );
+      break;
+    default:
+      payload = "0x";
+      break;
+  }
+
+  let nonce = await keyManager.callStatic.getNonce(reentrantSigner.address, 1);
+  let msgValue = 0;
+  let signature = await signLSP6ExecuteRelayCall(
+    keyManager,
+    nonce.toString(),
+    reentrantSigner.privateKey,
+    msgValue,
+    payload
+  );
+
+  await reentrancyRelayer.prepareRelayCall(signature, nonce, payload);
+};
+
+export const generateBatchRelayPayload = async (
+  reentrancyRelayer: RelayBatchReentrancy,
+  universalProfile: UniversalProfile,
+  keyManager: LSP6KeyManager,
+  reentrantSigner: Wallet,
+  payloadType: string,
+  testedAddress: string
+) => {
+  let payload: BytesLike;
+  switch (payloadType) {
+    case "TRANSFERVALUE":
+      payload = universalProfile.interface.encodeFunctionData(
+        "execute(uint256,address,uint256,bytes)",
+        [0, reentrancyRelayer.address, ethers.utils.parseEther("1"), "0x"]
+      );
+      break;
+    case "SETDATA":
+      payload = universalProfile.interface.encodeFunctionData(
+        "setData(bytes32,bytes)",
+        [
+          ethers.utils.keccak256(
+            ethers.utils.toUtf8Bytes("SomeRandomTextUsed")
+          ),
+          ethers.utils.hexlify(ethers.utils.toUtf8Bytes("SomeRandomTextUsed")),
+        ]
+      );
+      break;
+    case "ADDPERMISSIONS":
+      payload = universalProfile.interface.encodeFunctionData(
+        "setData(bytes32,bytes)",
+        [
+          ERC725YDataKeys.LSP6["AddressPermissions:Permissions"] +
+            testedAddress.substring(2),
+          ALL_PERMISSIONS,
+        ]
+      );
+      break;
+    case "CHANGEPERMISSIONS":
+      payload = universalProfile.interface.encodeFunctionData(
+        "setData(bytes32,bytes)",
+        [
+          ERC725YDataKeys.LSP6["AddressPermissions:Permissions"] +
+            testedAddress.substring(2),
+          "0x",
+        ]
+      );
+      break;
+    case "ADDUNIVERSALRECEIVERDELEGATE":
+      payload = universalProfile.interface.encodeFunctionData(
+        "setData(bytes32,bytes)",
+        [
+          ERC725YDataKeys.LSP1.LSP1UniversalReceiverDelegatePrefix +
+            ethers.utils
+              .keccak256(ethers.utils.toUtf8Bytes("RandomLSP1TypeId"))
+              .substring(2, 42),
+          testedAddress,
+        ]
+      );
+      break;
+    case "CHANGEUNIVERSALRECEIVERDELEGATE":
+      payload = universalProfile.interface.encodeFunctionData(
+        "setData(bytes32,bytes)",
+        [
+          ERC725YDataKeys.LSP1.LSP1UniversalReceiverDelegatePrefix +
+            ethers.utils
+              .keccak256(ethers.utils.toUtf8Bytes("RandomLSP1TypeId"))
+              .substring(2, 42),
+          "0x",
+        ]
+      );
+      break;
+    default:
+      payload = "0x";
+      break;
+  }
+
+  let nonce = await keyManager.callStatic.getNonce(reentrantSigner.address, 1);
+  let msgValue = 0;
+  let signature = await signLSP6ExecuteRelayCall(
+    keyManager,
+    nonce.toString(),
+    reentrantSigner.privateKey,
+    msgValue,
+    payload
+  );
+
+  await reentrancyRelayer.prepareRelayCall(
+    [signature],
+    [nonce],
+    [msgValue],
+    [payload]
+  );
+};
